Move sign-in and -out into route group 'auth'
