+++ frontend/src/app/(auth)/sign-up/page.client.tsx
