+++ frontend/src/app/(auth)/sign-up/page.tsx
@@ -1,5 +1,5 @@
 import {Section} from '@/components/Section'
-import {SignUpForm} from '@/app/sign-up/page.client'
+import {SignUpForm} from '@/app/(auth)/sign-up/page.client'
 
 export default function SignUpPage () {
     return (
