+++ frontend/src/app/@modal/(...)sign-up/page.tsx
@@ -1,6 +1,6 @@
 import {Modal, ModalActions} from '@/components/Modal'
-import {SignUpForm} from '@/app/sign-up/page.client'
 import {CloseModalButton} from '@/components/Modal.client'
+import {SignUpForm} from '@/app/(auth)/sign-up/page.client'
 
 
 export default function SignUpModal () {
